fold DC::Redis::Base module into DC::Redis::Truncation.
