--- lib/database_cleaner/redis/base.rb
@@ -1,41 +1,0 @@
-require 'database_cleaner/generic/base'
-
-module DatabaseCleaner
-  module Redis
-    def self.available_strategies
-      %i[truncation]
-    end
-
-    def self.default_strategy
-      available_strategies.first
-    end
-
-    module Base
-      include ::DatabaseCleaner::Generic::Base
-
-      def db=(desired_db)
-        @db = desired_db
-      end
-
-      def db
-        @db ||= :default
-      end
-
-      alias url db
-
-      private
-
-      def connection
-        @connection ||= begin
-          if url == :default
-            ::Redis.new
-          elsif db.is_a?(::Redis) # pass directly the connection
-            db
-          else
-            ::Redis.new(:url => url)
-          end
-        end
-      end
-    end
-  end
-end
