+++ spec/database_cleaner/redis/truncation_spec.rb
@@ -40,5 +40,24 @@ RSpec.describe DatabaseCleaner::Redis::Truncation do
       expect(@redis.get('Widget')).to eq '1'
     end
   end
-end
 
+  context "when passing url" do
+    it "should store my describe db" do
+      url = 'redis://localhost:6379/2'
+      subject.db = 'redis://localhost:6379/2'
+      expect(subject.db).to eq url
+    end
+  end
+
+  context "when passing connection" do
+    it "should store my describe db" do
+      connection = ::Redis.new :url => 'redis://localhost:6379/2'
+      subject.db = connection
+      expect(subject.db).to eq connection
+    end
+  end
+
+  it "should default to :default" do
+    expect(subject.db).to eq :default
+  end
+end
